+++ examples/tests/ModEM_PlotResponse.py
@@ -27,8 +27,7 @@ import os
 import os.path as op
 
 os.chdir(r'C:\Git\mtpy')
-from mtpy.modeling.modem import PlotResponse
-from mtpy.modeling.ModEM import Data
+from mtpy.modeling.ModEM import Data, PlotResponse
 
 #### Inputs ####
 wd = r'C:\Git\mtpy\examples\model_files\ModEM'
